+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/driveto/DriveToParams.java
@@ -3,7 +3,7 @@ package org.firstinspires.ftc.teamcode.driveto;
 import org.firstinspires.ftc.teamcode.utils.Heading;
 
 public class DriveToParams {
-    private static final double ROTATION_TARGET_RANGE = Heading.QUARTER_CIRCLE;
+    private static final float ROTATION_TARGET_RANGE = Heading.QUARTER_CIRCLE;
 
     // Administrative members
     public final DriveToListener parent;
@@ -13,9 +13,9 @@ public class DriveToParams {
 
     // Comparison data
     public DriveToComp comparator = DriveToComp.LESS;
-    public double limit = 0.0d;
-    public double limitRange = 0.0d; // Used in range, rotational, and PID comparators
-    public Double diffRange = null; // Used in PID comparators
+    public float limit = 0.0f;
+    public float limitRange = 0.0f; // Used in range, rotational, and PID comparators
+    public Float diffRange = null; // Used in PID comparators
     public boolean crossing = false; // Used in rotational comparators
 
     public DriveToParams(DriveToListener parent, Object reference) {
@@ -23,39 +23,39 @@ public class DriveToParams {
         this.reference = reference;
     }
 
-    public void lessThan(double limit) {
+    public void lessThan(float limit) {
         this.comparator = DriveToComp.LESS;
         this.limit = limit;
     }
 
     public void lessThan(int limit) {
-        lessThan((double) limit);
+        lessThan((float) limit);
     }
 
-    public void greaterThan(double limit) {
+    public void greaterThan(float limit) {
         this.comparator = DriveToComp.GREATER;
         this.limit = limit;
     }
 
     public void greaterThan(int limit) {
-        greaterThan((double) limit);
+        greaterThan((float) limit);
     }
 
-    public void rotationLess(double limit) {
+    public void rotationLess(float limit) {
         this.comparator = DriveToComp.ROTATION_LESS;
         this.limit = Heading.normalize(limit);
         this.limitRange = Heading.normalize(limit - ROTATION_TARGET_RANGE);
         this.crossing = this.limitRange > this.limit;
     }
 
-    public void rotationGreater(double limit) {
+    public void rotationGreater(float limit) {
         this.comparator = DriveToComp.ROTATION_GREATER;
         this.limit = Heading.normalize(limit);
         this.limitRange = Heading.normalize(limit + ROTATION_TARGET_RANGE);
         this.crossing = this.limitRange < this.limit;
     }
 
-    public void pid(double target, PIDParams params, double tolerance, Double diffTolerance) {
+    public void pid(float target, PIDParams params, float tolerance, Float diffTolerance) {
         this.comparator = DriveToComp.PID;
         this.pid = new PID(params);
         this.pid.setTarget(target);
@@ -66,7 +66,7 @@ public class DriveToParams {
         }
     }
 
-    public void rotationPid(double target, PIDParams params, double tolerance, Double diffTolerance) {
+    public void rotationPid(float target, PIDParams params, float tolerance, Float diffTolerance) {
         this.comparator = DriveToComp.ROTATION_PID;
         this.pid = new PID(params);
         this.pid.setTarget(target);
